requirements.txt
requests
beautifulsoup4
# pdf
pypdf>=5.6.0
reportlab
loguru
# docx
mammoth
python-docx
# excel
markitdown[xlsx]
# app
pydantic
pyyaml
# llm
openai
opencc-python-reimplemented
# api
fastapi
uvicorn
python-multipart
